solution/deploy/app/requirements.txt
streamlit
sqlalchemy
psycopg2-binary
requests
python-dotenv
numpy
tensorflow
